refactor(msteamsv2): move configuration types into msteamsv2 package
