+++ notify/msteamsv2/msteamsv2_test.go
@@ -31,7 +31,6 @@ import (
 
 	amcommoncfg "github.com/prometheus/alertmanager/config/common"
 
-	"github.com/prometheus/alertmanager/config"
 	"github.com/prometheus/alertmanager/notify"
 	"github.com/prometheus/alertmanager/notify/test"
 	"github.com/prometheus/alertmanager/types"
@@ -42,7 +41,7 @@ var testWebhookURL, _ = url.Parse("https://example.westeurope.logic.azure.com:44
 
 func TestMSTeamsV2Retry(t *testing.T) {
 	notifier, err := New(
-		&config.MSTeamsV2Config{
+		&MSTeamsV2Config{
 			WebhookURL: &amcommoncfg.SecretURL{URL: testWebhookURL},
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
 		},
@@ -75,7 +74,7 @@ func TestNotifier_Notify_WithReason(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			notifier, err := New(
-				&config.MSTeamsV2Config{
+				&MSTeamsV2Config{
 					WebhookURL: &amcommoncfg.SecretURL{URL: testWebhookURL},
 					HTTPConfig: &commoncfg.HTTPClientConfig{},
 				},
@@ -125,14 +124,14 @@ func TestMSTeamsV2Templating(t *testing.T) {
 
 	for _, tc := range []struct {
 		title string
-		cfg   *config.MSTeamsV2Config
+		cfg   *MSTeamsV2Config
 
 		retry  bool
 		errMsg string
 	}{
 		{
 			title: "full-blown message",
-			cfg: &config.MSTeamsV2Config{
+			cfg: &MSTeamsV2Config{
 				Title: `{{ template "msteams.default.title" . }}`,
 				Text:  `{{ template "msteams.default.text" . }}`,
 			},
@@ -140,14 +139,14 @@ func TestMSTeamsV2Templating(t *testing.T) {
 		},
 		{
 			title: "title with templating errors",
-			cfg: &config.MSTeamsV2Config{
+			cfg: &MSTeamsV2Config{
 				Title: "{{ ",
 			},
 			errMsg: "template: :1: unclosed action",
 		},
 		{
 			title: "message with templating errors",
-			cfg: &config.MSTeamsV2Config{
+			cfg: &MSTeamsV2Config{
 				Title: `{{ template "msteams.default.title" . }}`,
 				Text:  "{{ ",
 			},
@@ -191,7 +190,7 @@ func TestMSTeamsV2RedactedURL(t *testing.T) {
 
 	secret := "secret"
 	notifier, err := New(
-		&config.MSTeamsV2Config{
+		&MSTeamsV2Config{
 			WebhookURL: &amcommoncfg.SecretURL{URL: u},
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
 		},
@@ -213,7 +212,7 @@ func TestMSTeamsV2ReadingURLFromFile(t *testing.T) {
 	require.NoError(t, err, "writing to temp file failed")
 
 	notifier, err := New(
-		&config.MSTeamsV2Config{
+		&MSTeamsV2Config{
 			WebhookURLFile: f.Name(),
 			HTTPConfig:     &commoncfg.HTTPClientConfig{},
 		},
